requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
#conda==4.8.2
#conda-package-handling==1.6.0
cryptography==2.8
idna==2.8
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
requests==2.22.0
ruamel-yaml==0.15.87
six==1.14.0
tqdm==4.42.1
urllib3==1.25.8
PyGithub>=0.15
